Refactor, extract FarmScreen.js
